requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
lxml==4.5.0
matplotlib==2.2.3
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2019.3
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
XlsxWriter==1.2.8
psycopg2==2.8.4